+++ src/client/user/UserStatistics/UserAccuracyContainer/UserAccuracyContainer-nightmode.less
@@ -1,7 +1,7 @@
 @import (reference) '../../../styles/custom.less';
 
 .nightmode {
-  .UserStatisticContainer {
+  .UserAccuracyContainer {
     border: 1px solid @normal-grey;
     background-color: @grey-nero;
 
@@ -12,7 +12,7 @@
     }
   }
 
-  .UserStatisticContainer__period-item.border {
+  .UserAccuracyContainer__period-item.border {
     border-top: 1px solid @normal-grey;
     border-left: 1px solid @normal-grey;
   }
