+++ src/client/user/UserStatistics/UserAccuracyContainer/UserAccuracyContainer.less
@@ -1,6 +1,6 @@
 @import (reference) '../../../styles/custom.less';
 
-.UserStatisticContainer {
+.UserAccuracyContainer {
   display: flex;
   width: 100%;
   flex-direction: column;
@@ -32,10 +32,10 @@
   }
 }
 
-.UserStatisticContainer__period-item.border {
+.UserAccuracyContainer__period-item.border {
   width: 25%;
   border-top: 1px solid @white-gainsboro;
   border-left: 1px solid @white-gainsboro;
 }
 
-@import 'UserStatisticContainer-nightmode';
+@import 'UserAccuracyContainer-nightmode';
