+++ src/client/user/UserStatistics/UserStatistics.js
@@ -1,13 +1,13 @@
-import React, {useEffect, useState} from 'react';
-import {injectIntl} from 'react-intl';
-import {isEmpty} from 'lodash';
-import UserStatisticContainer from './UserStatisticContainer/UserStatisticContainer';
+import React, { useEffect, useState } from 'react';
+import { injectIntl } from 'react-intl';
+import { isEmpty } from 'lodash';
+import UserStatisticContainer from './UserAccuracyContainer/UserAccuracyContainer';
 import UserForecastInstruments from './UserForecastInstruments/UserForecastInstruments';
 import api from '../../../investarena/configApi/apiResources';
 import './UserStatistics.less';
 import UserInstrumentsTable from './UserInstrumentsTable/UserInstrumentsTable';
 
-const UserStatistics = ({match}) => {
+const UserStatistics = ({ match }) => {
   const [statAccuracyData, setStatAccuracyData] = useState({});
   const [sortOptions, setSortOptions] = useState({});
   const [statInstrumentsData, setStatInstrumentsData] = useState({});
@@ -15,7 +15,6 @@ const UserStatistics = ({match}) => {
   const parseOption = () => {
     const parsedOptions = {};
     if (!isEmpty(sortOptions)) {
-
       parsedOptions.sortDirection = sortOptions.isActive ? -1 : 1;
       parsedOptions.sortBy = sortOptions.currentItem;
     }
@@ -35,14 +34,14 @@ const UserStatistics = ({match}) => {
     <div className="UserStatistics">
       {!isEmpty(statAccuracyData) && (
         <React.Fragment>
-          <UserStatisticContainer accuracy={statAccuracyData} contentType={'forecast'}/>
-          <UserStatisticContainer accuracy={statAccuracyData} contentType={'profitability'}/>
+          <UserStatisticContainer accuracy={statAccuracyData} contentType={'forecast'} />
+          <UserStatisticContainer accuracy={statAccuracyData} contentType={'profitability'} />
         </React.Fragment>
       )}
       {!isEmpty(statInstrumentsData) && (
         <React.Fragment>
-          <UserForecastInstruments forecasts={statInstrumentsData}/>
-          <UserInstrumentsTable forecasts={statInstrumentsData} setSortOptions={setSortOptions}/>
+          <UserForecastInstruments forecasts={statInstrumentsData} />
+          <UserInstrumentsTable forecasts={statInstrumentsData} setSortOptions={setSortOptions} />
         </React.Fragment>
       )}
     </div>
