+++ src/investarena/api/Statistics.js
@@ -1,6 +1,6 @@
 import Base from './Base';
 import config from '../configApi/config';
-import {isEmpty} from 'lodash';
+import { isEmpty } from 'lodash';
 
 export default class Statistics extends Base {
   getUserStatistics(userName) {
@@ -10,7 +10,15 @@ export default class Statistics extends Base {
   }
   getUserInstrumentStatistics(userName, sortOptions) {
     return this.apiClient
-      .get(`${config.statistics.getUserInstrumentStatistics}${userName}?limit=200${!isEmpty(sortOptions) ? `&sortBy=${sortOptions.sortBy}&sortDirection=${sortOptions.sortDirection}`:''}`, {}, {})
+      .get(
+        `${config.statistics.getUserInstrumentStatistics}${userName}?limit=200${
+          !isEmpty(sortOptions)
+            ? `&sortBy=${sortOptions.sortBy}&sortDirection=${sortOptions.sortDirection}`
+            : ''
+        }`,
+        {},
+        {},
+      )
       .then(response => response);
   }
 }
